feat: allow subqueries inside lists of objects
All fields with collection: true and type: "object" are mapped as type: "nested" in ES.

When a field has mapping type "nested", we must use nested queries and nested aggregations for querying and statistics to work.

nested in ES is almost as if each element in a list is indexed as a separate document.

Implements #250