fix: indices was missing sometimes for array highlighting
The issue was that we always get the "old" unindexed highlights anyway. But finding the indexed highlights failed due to an assumption that either there cannot be any highlights on the level if there are inner hits. But sometimes there are both. Now that it works again, it replaces the "old" highlights and we only get the indexed highlights.